refactor(egress): extract serviceOf to engine/service-of.ts (break import cycle)
egress-record.ts imported serviceOf from engine/executor.ts, which imports the
egress record builder back — a dependency cycle that audit:boundaries rejects.
Move serviceOf into its own zero-dep module both consume. executor.ts stays
flagship-clean (the function's branch coverage moves with it; the flagship
serviceOf test now exercises service-of.ts directly).
